separate module from vit class
